[fix] Fix collision glitches and change collision resolution
Fix a bug collision where the ball can be stuck on the Vaus. Fix a
glitch where the ball can be stuck on the left, right and top pipe.
Change collision resolution between the ball and the blocks. Change ball
physics when hitting the Vaus. Center origin of the blocks.